# tests/hosts/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "ActionResult": {
        "additionalProperties": false,
        "description": "Safe result of a completed provider mutation.",
        "properties": {
          "action_id": {
            "pattern": "^[a-z][a-z0-9_]{0,63}$",
            "title": "Action Id",
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "title": "Completed At",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/ContentRef"
          }
        },
        "required": [
          "action_id",
          "ref",
          "idempotency_key",
          "completed_at"
        ],
        "title": "ActionResult",
        "type": "object"
      },
      "ActionResultResponse": {
        "additionalProperties": false,
        "properties": {
          "result": {
            "$ref": "#/components/schemas/ActionResult"
          }
        },
        "required": [
          "result"
        ],
        "title": "ActionResultResponse",
        "type": "object"
      },
...
      "BaseModel": {
        "properties": {},
        "title": "BaseModel",
        "type": "object"
      },
...
      "ConfirmActionRequest": {
        "additionalProperties": false,
        "properties": {
          "pending_action_id": {
            "pattern": "^pending_[0-9a-f]{32}$",
            "title": "Pending Action Id",
            "type": "string"
          },
          "user_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "pending_action_id",
          "user_id"
        ],
        "title": "ConfirmActionRequest",
        "type": "object"
      },
...
      "ContentKind": {
        "additionalProperties": false,
        "description": "Provider-declared stable content-kind identity.",
        "properties": {
          "value": {
            "pattern": "^[a-z][a-z0-9_-]{0,63}$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "ContentKind",
        "type": "object"
      },
...
      "ContentPreview": {
        "additionalProperties": false,
        "description": "Small read/tool preview; not a recommendation or card schema.",
        "properties": {
          "creator_label": {
            "anyOf": [
              {
                "maxLength": 300,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Creator Label"
          },
          "provenance": {
            "$ref": "#/components/schemas/ProjectionProvenance"
          },
          "ref": {
            "$ref": "#/components/schemas/ContentRef"
          },
          "source_timestamp": {
            "format": "date-time",
            "title": "Source Timestamp",
            "type": "string"
          },
          "summary": {
            "maxLength": 4000,
            "title": "Summary",
            "type": "string"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "title",
          "summary",
          "source_timestamp",
          "provenance"
        ],
        "title": "ContentPreview",
        "type": "object"
      },
      "ContentRef": {
        "additionalProperties": false,
        "description": "Canonical cross-process identity for one provider-native record.\n\nURL normalization remains provider-owned. This layer validates and stores\nthe provider's canonical HTTP(S) result without rewriting it.",
        "properties": {
          "canonical_url": {
            "maxLength": 2048,
            "pattern": "^https?://[^\\s]+$",
            "title": "Canonical Url",
            "type": "string"
          },
          "content_kind": {
            "$ref": "#/components/schemas/ContentKind"
          },
          "provider_content_id": {
            "maxLength": 512,
            "minLength": 1,
            "title": "Provider Content Id",
            "type": "string"
          },
          "provider_id": {
            "$ref": "#/components/schemas/ProviderId"
          }
        },
        "required": [
          "provider_id",
          "content_kind",
          "provider_content_id",
          "canonical_url"
        ],
        "title": "ContentRef",
        "type": "object"
      },
      "ContentResponse": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "$ref": "#/components/schemas/NativeContent"
          }
        },
        "required": [
          "content"
        ],
        "title": "ContentResponse",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "validation",
          "unauthorized",
          "forbidden",
          "unavailable_capability",
          "conflict",
          "rate_limit",
          "temporary_failure",
          "not_found",
          "method_not_allowed"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "message": {
            "maxLength": 500,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "NativeContent": {
        "additionalProperties": false,
        "description": "Validated provider-native record at the heterogeneous registry boundary.\n\nProviders must validate external JSON into their own Pydantic model before\nconstructing this envelope. Untyped mappings are rejected.",
        "properties": {
          "payload": {
            "$ref": "#/components/schemas/BaseModel"
          },
          "ref": {
            "$ref": "#/components/schemas/ContentRef"
          },
          "schema_version": {
            "minimum": 1.0,
            "title": "Schema Version",
            "type": "integer"
          }
        },
        "required": [
          "ref",
          "schema_version",
          "payload"
        ],
        "title": "NativeContent",
        "type": "object"
      },
...
      "PendingAction": {
        "additionalProperties": false,
        "description": "Safe immutable confirmation record; never stores credentials or executable payloads.",
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "action_id": {
            "pattern": "^[a-z][a-z0-9_]{0,63}$",
            "title": "Action Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "pending_action_id": {
            "pattern": "^pending_[0-9a-f]{32}$",
            "title": "Pending Action Id",
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/ContentRef"
          },
          "result": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ActionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "safe_preview": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Safe Preview",
            "type": "string"
          },
          "user_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "pending_action_id",
          "idempotency_key",
          "action_id",
          "ref",
          "user_id",
          "safe_preview",
          "created_at",
          "expires_at"
        ],
        "title": "PendingAction",
        "type": "object"
      },
      "PendingActionResponse": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "$ref": "#/components/schemas/PendingAction"
          }
        },
        "required": [
          "action"
        ],
        "title": "PendingActionResponse",
        "type": "object"
      },
...
      "ProjectionProvenance": {
        "additionalProperties": false,
        "description": "Native source identity and projection time.",
        "properties": {
          "native_schema_version": {
            "minimum": 1.0,
            "title": "Native Schema Version",
            "type": "integer"
          },
          "projected_at": {
            "format": "date-time",
            "title": "Projected At",
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/ContentRef"
          }
        },
        "required": [
          "ref",
          "native_schema_version",
          "projected_at"
        ],
        "title": "ProjectionProvenance",
        "type": "object"
      },
      "ProposeActionRequest": {
        "additionalProperties": false,
        "properties": {
          "account_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "action_id": {
            "pattern": "^[a-z][a-z0-9_]{0,63}$",
            "title": "Action Id",
            "type": "string"
          },
          "expires_in_seconds": {
            "default": 300,
            "maximum": 3600.0,
            "minimum": 1.0,
            "title": "Expires In Seconds",
            "type": "integer"
          },
          "idempotency_key": {
            "maxLength": 200,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "ref": {
            "$ref": "#/components/schemas/ContentRef"
          },
          "safe_preview": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Safe Preview",
            "type": "string"
          },
          "user_id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "idempotency_key",
          "action_id",
          "ref",
          "user_id",
          "safe_preview"
        ],
        "title": "ProposeActionRequest",
        "type": "object"
      },
...
      "ProviderId": {
        "additionalProperties": false,
        "description": "Stable lower-case provider identity; never a display name.",
        "properties": {
          "value": {
            "pattern": "^[a-z][a-z0-9_-]{0,63}$",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "ProviderId",
        "type": "object"
      },
...
      "SearchResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContentPreview"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "SearchResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "OpenBiliClaw API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/content/actions/confirm": {
      "post": {
        "operationId": "confirm_v1_content_actions_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Confirm",
        "tags": [
          "content"
        ]
      }
    },
    "/v1/content/actions/propose": {
      "post": {
        "operationId": "propose_v1_content_actions_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProposeActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Propose",
        "tags": [
          "content"
        ]
      }
    },
    "/v1/content/search": {
      "get": {
        "operationId": "search_v1_content_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "provider_id",
            "required": true,
            "schema": {
              "title": "Provider Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Search",
        "tags": [
          "content"
        ]
      }
    },
    "/v1/content/{reference}": {
      "get": {
        "operationId": "detail_v1_content__reference__get",
        "parameters": [
          {
            "in": "path",
            "name": "reference",
            "required": true,
            "schema": {
              "title": "Reference",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "405": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "method_not_allowed"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "conflict"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "rate_limit"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "temporary_failure"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unavailable capability or temporary timeout"
          }
        },
        "summary": "Detail",
        "tags": [
          "content"
        ]
      }
    },
...
  }
}